URL fix
Looks like they revamped the linked site and clobbered the old url. You appear to only be able to view the pdf version or download the zipped html. What do you think about having a "readings" area in the repository that contains the full copies of such freely licensed works?